# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Learning Server API",
    "description": "Comprehensive API documentation for the E-Learning platform server. This API provides endpoints for user management, course management, authentication, payment processing, and administrative functions.",
    "version": "1.0.0",
    "contact": {
      "name": "E-Learning API Support",
      "email": "support@elearning.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.elearning.com",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Include in the 'token' header."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["message"]
      },
...
    }
  },
  "paths": {
...
    "/api/create-payment-link": {
      "post": {
        "tags": ["Payment"],
        "summary": "Create payment link",
        "description": "Create a payment link for course purchase",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Payment amount",
                    "example": 1000
                  },
                  "description": {
                    "type": "string",
                    "description": "Payment description",
                    "example": "Course purchase"
                  }
                },
                "required": ["amount", "description"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checkoutUrl": {
                      "type": "string",
                      "format": "uri",
                      "description": "Payment checkout URL"
                    },
                    "orderCode": {
                      "type": "integer",
                      "description": "Generated order code"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receive-hook": {
      "post": {
        "tags": ["Payment"],
        "summary": "Payment webhook",
        "description": "Receive payment webhook notifications",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Webhook payload from payment provider"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook received successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Echo of the received webhook data"
                }
              }
            }
          }
        }
      }
    },
...
  }
}